requirements_install.txt
Django==1.8.2
Fabric==1.10.2
Fuzzy==1.0
Genshi==0.6
Pillow==2.5.1
SQLAlchemy==0.9.9
astroid==1.0.1
cx-Oracle==5.1.1
djangorestframework==3.1.3
dt-mongodb-proxy==1.0
dt-mrg-addons==0.4
dt20db==0.2.10-103
dt20msg==0.19
dtconfig==1.1.1
dtconfig-data==1.1.17
dtcrypto==1.1.2
dtplatform==2.13.117-34
ecdsa==0.11
html5lib==0.95
jsonpickle==0.9.2
logilab-common==0.60.0
lxml==3.2.1
paramiko==1.15.2
pyPdf==1.13
pycrypto==2.6
pymongo==2.6.3
pymqi==1.3
python-dateutil==2.1
python-gnupg==0.3.0
pytz==2012j
redis==2.9.1
redis-collections==0.1.5
reportlab==2.6
requests==1.1.0
simplejson==2.6.0
simplepath==0.1.1
six==1.9.0
tzlocal==1.1.1
wsgiref==0.1.2
xhtml2pdf==0.0.5
zope.component==4.0.2
zope.event==4.0.2
zope.hookable==4.0.1
zope.interface==4.0.1
